pyproject.toml
[project]
name = "n8n-art-architect"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mcp>=1.11.0",
    "openai>=1.90.0",
    "openpipe-art==0.4.11",
    "openpipe-art[backend]==0.4.11 ; sys_platform == 'linux'",
    "pydantic>=2.12.4",
    "python-dotenv>=1.2.1",
    "tenacity>=9.1.2",
    "triton==3.3.0 ; sys_platform == 'linux'",
    "vllm==0.9.2 ; sys_platform == 'linux'",
    "wandb>=0.21.0",
]
